+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/ui/settings/GlobsConfigurable.kt
@@ -2,7 +2,6 @@ package com.github.blarc.gitlab.template.lint.plugin.ui.settings
 
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintBundle.message
 import com.github.blarc.gitlab.template.lint.plugin.extensions.createColumn
-import com.github.blarc.gitlab.template.lint.plugin.extensions.replaceAt
 import com.github.blarc.gitlab.template.lint.plugin.extensions.reportBugLink
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.intellij.openapi.options.BoundConfigurable
@@ -17,89 +16,118 @@ import com.intellij.util.ui.ListTableModel
 import javax.swing.ListSelectionModel.SINGLE_SELECTION
 
 class GlobsConfigurable : BoundConfigurable(message("settings.globs.group.title")) {
-    private var globs = AppSettings.instance.gitlabLintGlobStrings
+    private var inclusionGlobs = AppSettings.instance.gitlabLintGlobStrings.toMutableList()
+    private var exclusionGlobs = AppSettings.instance.exclusionGlobs.toMutableList()
 
-    private val tableModel = createTableModel()
+    private val inclusionTableModel = createTableModel(inclusionGlobs)
+    private val exclusionTableModel = createTableModel(exclusionGlobs)
 
-    private val table = TableView(tableModel).apply {
+    private val inclusionTable = TableView(inclusionTableModel).apply {
+        setShowColumns(false)
+        setSelectionMode(SINGLE_SELECTION)
+    }
+
+    private val exclusionTable = TableView(exclusionTableModel).apply {
         setShowColumns(false)
         setSelectionMode(SINGLE_SELECTION)
     }
 
     override fun createPanel() = panel {
         row {
+            label(message("settings.globs.inclusions.title"))
+        }
+        row {
             cell(
-                ToolbarDecorator.createDecorator(table)
-                    .setAddAction{ addGlob() }
-                    .setEditAction{ editGlob() }
-                    .setRemoveAction{ removeGlob() }
+                ToolbarDecorator.createDecorator(inclusionTable)
+                    .setAddAction { addGlob(inclusionGlobs) }
+                    .setEditAction { editGlob(inclusionTable, inclusionGlobs) }
+                    .setRemoveAction { removeGlob(inclusionTable, inclusionGlobs) }
                     .disableUpAction()
                     .disableDownAction()
                     .createPanel()
             ).align(Align.FILL)
-                .comment(message("settings.globs.comment"))
         }.resizableRow()
+        row {
+            label(message("settings.globs.exclusions.title"))
+        }
+        row {
+            cell(
+                ToolbarDecorator.createDecorator(exclusionTable)
+                    .setAddAction { addGlob(exclusionGlobs) }
+                    .setEditAction { editGlob(exclusionTable, exclusionGlobs) }
+                    .setRemoveAction { removeGlob(exclusionTable, exclusionGlobs) }
+                    .disableUpAction()
+                    .disableDownAction()
+                    .createPanel()
+            ).align(Align.FILL)
+        }.resizableRow()
+        row {
+            comment(message("settings.globs.comment"))
+        }
         row {
             reportBugLink()
         }
     }
 
-
-    private fun createTableModel(): ListTableModel<String> = ListTableModel(
+    private fun createTableModel(globs: List<String>): ListTableModel<String> = ListTableModel(
         arrayOf(
             createColumn<String>("Glob") { glob -> glob }
         ),
         globs
     )
 
-    private fun addGlob() {
+    private fun addGlob(globs: MutableList<String>) {
         val dialog = GlobDialog()
 
         if (dialog.showAndGet()) {
-            globs = globs.plus(dialog.glob)
+            globs.add(dialog.glob)
             refreshTableModel()
         }
     }
 
-    private fun removeGlob() {
-        val row = table.selectedObject ?: return
-
-        globs = globs.minus(row)
+    private fun removeGlob(table: TableView<String>, globs: MutableList<String>) {
+        val glob = table.selectedObject ?: return
+        globs.remove(glob)
         refreshTableModel()
     }
 
-    private fun editGlob() {
-        val row = table.selectedObject ?: return
+    private fun editGlob(table: TableView<String>, globs: MutableList<String>) {
+        val glob = table.selectedObject ?: return
 
-        val dialog = GlobDialog(row)
+        val dialog = GlobDialog(glob)
         if (dialog.showAndGet()) {
-            globs = if (dialog.glob.isEmpty()) {
-                globs.minus(row)
+            if (dialog.glob.isEmpty()) {
+                globs.remove(glob)
             } else {
-                globs.replaceAt(table.selectedRow, dialog.glob)
+                globs.remove(glob)
+                globs.add(dialog.glob)
             }
             refreshTableModel()
         }
     }
 
     private fun refreshTableModel() {
-        tableModel.items = globs
+        inclusionTableModel.items = inclusionGlobs
+        exclusionTableModel.items = exclusionGlobs
     }
 
     override fun reset() {
         super.reset()
-
-        globs = AppSettings.instance.gitlabLintGlobStrings
+        inclusionGlobs = AppSettings.instance.gitlabLintGlobStrings.toMutableList()
+        exclusionGlobs = AppSettings.instance.exclusionGlobs.toMutableList()
         refreshTableModel()
     }
 
     override fun isModified(): Boolean {
-        return super.isModified() || globs != AppSettings.instance.gitlabLintGlobStrings
+        return super.isModified() ||
+                inclusionGlobs != AppSettings.instance.gitlabLintGlobStrings ||
+                exclusionGlobs != AppSettings.instance.exclusionGlobs
     }
 
     override fun apply() {
         super.apply()
-        AppSettings.instance.gitlabLintGlobStrings = globs
+        AppSettings.instance.gitlabLintGlobStrings = inclusionGlobs
+        AppSettings.instance.exclusionGlobs = exclusionGlobs
     }
 }
 
@@ -120,7 +148,7 @@ private class GlobDialog(var glob: String = "") : DialogWrapper(false) {
             label(message("settings.glob.dialog.label"))
             textField()
                 .text(glob)
-                .bindText({glob}, {glob = it})
+                .bindText({ glob }, { glob = it })
                 .focused()
         }
     }