+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Exclusion globs for excluding files from linting.
 
 ## [1.6.0] - 2023-03-31